refactor: move provider_metrics_invariant_met into single function
