+++ rust/main/utils/run-locally/src/sealevel/termination_invariants.rs
@@ -5,8 +5,8 @@ use maplit::hashmap;
 use crate::{
     config::Config,
     invariants::{
-        relayer_termination_invariants_met, scraper_termination_invariants_met,
-        RelayerTerminationInvariantParams,
+        provider_metrics_invariant_met, relayer_termination_invariants_met,
+        scraper_termination_invariants_met, RelayerTerminationInvariantParams,
     },
     logging::log,
     sealevel::{solana::*, SOL_MESSAGES_EXPECTED, SOL_MESSAGES_WITH_NON_MATCHING_IGP},
@@ -63,23 +63,14 @@ pub fn termination_invariants_met(
         return Ok(false);
     }
 
-    if !request_metric_invariant_met(total_messages_expected)? {
+    if !provider_metrics_invariant_met(
+        RELAYER_METRICS_PORT,
+        total_messages_expected,
+        &hashmap! {"chain" => "sealeveltest1", "status" => "success"},
+    )? {
         return Ok(false);
     }
 
     log!("Termination invariants have been meet");
-    Ok(true)
-}
-
-pub fn request_metric_invariant_met(expected_request_count: u32) -> eyre::Result<bool> {
-    let request_count = fetch_metric(
-        RELAYER_METRICS_PORT,
-        "hyperlane_request_count",
-        &hashmap! {"chain" => "sealeveltest1", "status" => "success"},
-    )?
-    .iter()
-    .sum::<u32>();
-
-    assert!(request_count > expected_request_count);
     Ok(true)
 }
